% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_products.R
\name{create_image_product}
\alias{create_image_product}
\title{create_image_collection}
\usage{
create_image_product(productID = "CGIAR/SRTM90_V4",
  productName = "srtm", spatialReducer = "mean", resolution = 3000)
}
\arguments{
\item{productID}{Strong of the Image/ImageCollection ID found in Earth Engine Data Explorer}

\item{productName}{A name for the data product specified by the user.}

\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{temporalReducer}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}

\item{yearIntervall}{A path to a local file or a name of a already uploaded to earth engine}
}
\value{
depend on output
}
\description{
Climate Hazards Group InfraRed Precipitation with Station data (CHIRPS) is a 30+ year quasi-global rainfall dataset. CHIRPS incorporates 0.05° resolution satellite imagery with in-situ station data to create gridded rainfall time series for trend analysis and seasonal drought monitoring.
}
