% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{data_jrc_distanceToWater}
\alias{data_jrc_distanceToWater}
\title{jrc_distanceToWater}
\usage{
data_jrc_distanceToWater(spatialReducer = "mean", temporalReducer = "mean",
  yearIntervall = c(2000, 2002))
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{temporalReducer}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}

\item{yearIntervall}{A path to a local file or a name of a already uploaded to earth engine}
}
\value{
depend on output
}
\description{
jrc_distanceToWater
}
