% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{data_oxford_friction}
\alias{data_oxford_friction}
\title{oxford_friction}
\usage{
data_oxford_friction(spatialReducer = "mean")
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}
}
\value{
depend on output
}
\description{
oxford_friction
}
