% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{data_srtm_slope}
\alias{data_srtm_slope}
\title{srtm_slope}
\usage{
data_srtm_slope(spatialReducer = "mean")
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}
}
\value{
depend on output
}
\description{
srtm_slope
}
