% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{delete_credentials}
\alias{delete_credentials}
\title{deletes credentials to re initialize}
\usage{
delete_credentials(credentials = c("gd-credentials.rds", "credentials",
  "ft_credentials.json"))
}
\description{
deletes credentials to re initialize
}
