% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{download_data}
\usage{
download_data(info, path = getwd(), clear = T)
}
\arguments{
\item{info}{Output of the get_data function.}

\item{path}{The local path where the file should be stored, default is working directory.}

\item{clear}{If the file should be removed from Google Drive after the download.}
}
\value{
nothing
}
\description{
download_data
}
