% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{eeProduct_modis_nonTreeVegetation}
\alias{eeProduct_modis_nonTreeVegetation}
\title{modis_nonTreeVegetation}
\usage{
eeProduct_modis_nonTreeVegetation(spatialReducer = "mean",
  temporalReducer = "mean", yearIntervall = c(2000, 2002))
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{temporalReducer}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}

\item{yearIntervall}{A path to a local file or a name of a already uploaded to earth engine}
}
\value{
depend on output
}
\description{
modis_nonTreeVegetation
}
