% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{eeProduct_srtm_slope}
\alias{eeProduct_srtm_slope}
\title{srtm_slope}
\usage{
eeProduct_srtm_slope(spatialReducer = "mean")
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}
}
\value{
depend on output
}
\description{
srtm_slope
}
