% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_file}
\alias{find_file}
\title{find a file in a speciefied subdirectory}
\usage{
find_file(filename, root_dir)
}
\arguments{
\item{filename}{the name of the file to search, with extension}

\item{root_dir}{the initial directory the to search}
}
\value{
the given file with the path
}
\description{
find a file in a speciefied subdirectory
}
