% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_folder}
\alias{find_folder}
\title{find a folder in a speciefied subdirectory}
\usage{
find_folder(foldername, root_dir)
}
\arguments{
\item{foldername}{the name of the folder to search}

\item{root_dir}{the initial directory the to search}
}
\value{
the given folder with the path
}
\description{
find a folder in a speciefied subdirectory
}
