% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_in_one}
\alias{get_all_in_one}
\title{get_all_in_one}
\usage{
get_all_in_one(month_start = 1, month_end = 12, year_start = 2000,
  year_end = 2000, asset_path = "users/JesJehle/Strips")
}
\arguments{
\item{month_start}{Integer for month start}

\item{month_end}{Integer for month (1 - 12)}

\item{year_start}{Integer (1981 - 2017)}

\item{year_end}{Integer (1981 - 2017)}

\item{asset_path}{path of asset in GEE}
}
\value{
data.frame with data
}
\description{
get_all_in_one
}
