% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_chirps}
\alias{get_chirps}
\title{get_chirps}
\usage{
get_chirps(month_start = 1, month_end = 12, year_start = 2000,
  year_end = 2000, asset_path = "users/JesJehle/Strips")
}
\arguments{
\item{month_start}{Integer for month start}

\item{month_end}{Integer for month (1 - 12)}

\item{year_start}{Integer (1981 - 2017)}
}
\value{
data.frame with ID and precipitation
}
\description{
get_chirps
}
