% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{get_credential_path}
\alias{get_credential_path}
\title{Find path to specified credentials folder}
\usage{
get_credential_path()
}
\value{
path to credentials folder
}
\description{
Find path to specified credentials folder
}
