% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_credential_root}
\alias{get_credential_root}
\title{Find path to specified credentials folder}
\usage{
get_credential_root()
}
\value{
path to credentials folder
}
\description{
Find path to specified credentials folder
}
