% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(timeIntervall = c(2000, 2000), assetPath = NULL,
  spatialReducer = "mean", outputFormat = "GeoJSON", resolution = 100,
  name = "example", products = list(c("chirps_precipitation", "mean"),
  c("jrc_permanentWater", "mean"), c("modis_treeCover", "mean"),
  c("modis_nonTreeVegetation", "mean"), c("modis_nonVegetated", "mean"),
  c("srtm_elevation", NULL), c("srtm_slope", NULL), c("modis_quality", "mean"),
  c("oxford_friction", NULL), c("oxford_accessibility", NULL)))
}
\arguments{
\item{timeIntervall}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}

\item{assetPath}{A string path to earth engine asset}

\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{outputFormat}{A string specifying the output format: CSV, GeoJSON, KML or KMZ}

\item{resolution}{Resolution of the dataproducts. modis = 250}

\item{products}{list of vectors as list(c(dataproduct, timeReducer)...)}

\item{timeReducer}{Reducer to aggregate data over time, can be: mean, median or mode, sum, min, max}
}
\value{
depend on output
}
\description{
get_data
}
