% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_old}
\alias{get_data_old}
\title{get_data}
\usage{
get_data_old(timeperiod = c(2000, 2005), produkt_list = c("precipitation",
  "surface_water", "treecover", "slope"),
  spatial_grid = "users/JesJehle/Strips")
}
\arguments{
\item{timeperiod}{Vector with first and last year of the timeperiod if interest}

\item{produkt_list}{Vector of available data produkts}

\item{spatial_grid}{Path of the desired spatial grid in the google earth engine Assets}
}
\value{
data.frame with selected data products
}
\description{
get_data
}
