% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_data_one_by_one}
\alias{get_data_one_by_one}
\title{get_data_one_by_one}
\usage{
get_data_one_by_one(target = NULL, outputFormat = "GeoJSON",
  resolution = 100, products = list(data_jrc_distanceToWater()))
}
\arguments{
\item{target}{A path to a local file or a name of a already uploaded to earth engine}

\item{outputFormat}{A string specifying the output format: CSV, GeoJSON, KML or KMZ}

\item{resolution}{Resolution of the dataproducts. modis = 250}

\item{products}{list of vectors as list(c(dataproduct, timeReducer)...)}

\item{timeReducer}{Reducer to aggregate data over time, can be: mean, median or mode, sum, min, max}

\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{timeIntervall}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}
}
\value{
depend on output
}
\description{
get_data_one_by_one
}
