% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_data_profile}
\alias{get_data_profile}
\title{get_data_profile_version}
\usage{
get_data_profile(timeIntervall = c(2000, 2000), assetPath = NULL,
  spatialReducer = "mean", outputFormat = "GeoJSON", resolution = 100,
  numPolygons = 10000, name = "example",
  products = list(c("chirps_precipitation", "mean"), c("jrc_distanceToWater",
  "mode"), c("modis_treeCover", "mean"), c("modis_nonTreeVegetation", "mean"),
  c("modis_nonVegetated", "mean"), c("srtm_elevation", NA), c("srtm_slope", NA),
  c("modis_quality", "mean"), c("oxford_friction", NA),
  c("oxford_accessibility", NA)))
}
\arguments{
\item{timeIntervall}{Integers to spedify the beginning and end of timeperiod to reduce over as c(yearStart, yearEnd).}

\item{assetPath}{A string path to earth engine asset}

\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}

\item{outputFormat}{A string specifying the output format: CSV, GeoJSON, KML or KMZ}

\item{resolution}{Resolution of the dataproducts. modis = 250}

\item{products}{list of vectors as list(c(dataproduct, timeReducer)...)}

\item{timeReducer}{Reducer to aggregate data over time, can be: mean, median or mode, sum, min, max}
}
\value{
depend on output
}
\description{
get_data_profile_version
}
