% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_data_test}
\alias{get_data_test}
\title{get_data_test}
\usage{
get_data_test(year_start = 2000, year_end = 2000, time_reducer = "mean",
  asset_path = "users/JesJehle/Strips", output = "GeoJSON",
  spatial_reducer = "mean", scale = 100, test = "no_size_test",
  numPolygons = 10000, name = "example", export = F,
  products = list("chirps_precipitation", "jrc_permanentWater",
  "modis_treeCover", "modis_nonTreeVegetation", "modis_nonVegetated",
  "srtm_elevation", "srtm_slope", "modis_quality"))
}
\arguments{
\item{year_start}{asd}

\item{year_end}{asd}

\item{time_reducer}{asd}

\item{asset_path}{}

\item{output}{A String specifying the output format: CSV, GeoJSON, KML or KMZ}

\item{spatial_reducer}{as}

\item{scale}{as}

\item{test}{what test to perform, "size_test"}

\item{numPolygons}{if test is "size_test", numPolygons sets number of polygons to test}

\item{products}{list("chirps_precipitation","jrc_permanentWater","modis_treeCover","modis_nonTreeVegetation","modis_nonVegetated","srtm_elevation",srtm_slope", "modis_quality")}

\item{raster}{perPolygon = list(raster = T, asset_id), raster = list(raster = T, extend, resolution), rasterAsDataFrame = list(rasterAsDataFrame = T, extand, resolution)}
}
\value{
depends on target
}
\description{
get_data_test
}
