% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_ft_id}
\alias{get_ft_id}
\title{get_ft_id}
\usage{
get_ft_id(ft_name)
}
\arguments{
\item{ft_name}{name of the fusiontable}
}
\value{
fusiontable ID or NA of no fusiontable with given name
}
\description{
get_ft_id
}
