% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{get_ft_id_gd}
\alias{get_ft_id_gd}
\title{get_ft_id_gd extracts fusion table ID}
\usage{
get_ft_id_gd(ft_name)
}
\arguments{
\item{ft_name}{Name of fusion table in google drive}
}
\description{
get_ft_id_gd extracts fusion table ID
}
