% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_glcf_treecover}
\alias{get_glcf_treecover}
\title{get_glcf_treecover}
\usage{
get_glcf_treecover(year_start = 1985, year_end = 2000,
  asset_path = "users/JesJehle/Strips")
}
\arguments{
\item{year_start}{Integer (2000, 2005, 2010)}

\item{year_end}{Integer (2000, 2005, 2010)}
}
\value{
data.frame with ID, percent treecover and uncertainty
}
\description{
get_glcf_treecover
}
