% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_jrc_year}
\alias{get_jrc_year}
\title{get_jrc_year}
\usage{
get_jrc_year(year_start = 1985, year_end = 2000,
  asset_path = "users/JesJehle/Strips")
}
\arguments{
\item{year_start}{Integer (1981 - 2017)}

\item{year_end}{Integer (1981 - 2017)}
}
\value{
data.frame with ID and mode of precipitation [mm] over the given interval
}
\description{
get_jrc_year
}
