% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_modis_treecover}
\alias{get_modis_treecover}
\title{get_modis_treecover}
\usage{
get_modis_treecover(year_start = 2000, year_end = 2001,
  asset_path = "users/JesJehle/Strips")
}
\arguments{
\item{year_start}{Integer (2000-2015)}

\item{year_end}{Integer (2000-2015)}
}
\value{
data.frame with ID, percent treecover percent non treecover percent non cover and quality
}
\description{
get_modis_treecover
}
