% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{get_name_from_path}
\alias{get_name_from_path}
\title{get_name_from_path}
\usage{
get_name_from_path(path)
}
\arguments{
\item{path}{A file path}
}
\value{
basename without extension
}
\description{
get_name_from_path
}
