% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_product_info}
\alias{get_product_info}
\title{get_product_info}
\usage{
get_product_info(prodct_id, path)
}
\arguments{
\item{path}{String, specifieng the path for the medata}

\item{product_id}{String of the product id}
}
\description{
Retrives Metadata of data product
}
