% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_sys}
\alias{get_sys}
\title{test sys connection}
\usage{
get_sys(month_start = 1, month_end = 5, year_start = 2000,
  year_end = 2000, asset_path = "users/JesJehle/Strips")
}
\value{
sys arguments
}
\description{
test sys connection
}
