% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_temp_path}
\alias{get_temp_path}
\title{get_temp_path}
\usage{
get_temp_path(create = T)
}
\arguments{
\item{create}{logical weather to create a new folder}
}
\description{
creates folder and returns path for the storage of local temp files, if folder alreday exists it gets deleted and new created.
}
