% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{import_data}
\alias{import_data}
\title{import data}
\usage{
import_data(productList, verbose = T, files_dir, clean = T)
}
\arguments{
\item{productList}{List of products files produced in the ee_grab function}

\item{verbose}{If true, messages reporting the processing state are printed.}
}
\value{
nothing
}
\description{
import data
}
