% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{initialize_gee2r}
\alias{initialize_gee2r}
\title{initialize gee2r package by installing the ee python library with all dependencies and creat earth engine credetiales.}
\usage{
initialize_gee2r()
}
\description{
initialize gee2r package by installing the ee python library with all dependencies and creat earth engine credetiales.
}
