% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_ee_dependencies}
\alias{install_ee_dependencies}
\title{The function installs python dependencies}
\usage{
install_ee_dependencies(conda_env_name)
}
\description{
The function installs python dependencies
}
