% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_type}
\alias{is_type}
\title{is_type}
\usage{
is_type(param, type)
}
\arguments{
\item{param}{the parameter to test}

\item{type}{the required type of the parameter}
}
\description{
test of param is of type type and raises an appropriate error
}
