% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{JsonToShapefile}
\alias{JsonToShapefile}
\title{jsonToShapefile}
\usage{
JsonToShapefile(json_path, shapefile_path)
}
\arguments{
\item{shapefile_path}{path and name  as String of the wanted shapefile}

\item{json_paht}{path as String to the json file}
}
\value{
ESRI Shapefile
}
\description{
jsonToShapefile
}
