% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{request_data}
\usage{
request_data(product_info, target_id, verbose = T, test = F)
}
\arguments{
\item{product_info}{list object created by ee_product functions}

\item{target_id}{String of fusion table id created by upload_data()}
}
\value{
ee_responses for each correctly exported data product
}
\description{
Starts processing on earth engine retrieves info from data product
}
