% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{run_gd_auth}
\alias{run_gd_auth}
\title{Runs google drive authorisation via googledrive::drive_auth()}
\usage{
run_gd_auth(credential_path, credential_name = ".httr-oauth")
}
\description{
Runs google drive authorisation via googledrive::drive_auth()
}
