% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{run_gd_auth_old}
\alias{run_gd_auth_old}
\title{old Runs google drive authorisation via googledrive::drive_auth()}
\usage{
run_gd_auth_old(credential_path, credential_name = ".httr-oauth")
}
\description{
old Runs google drive authorisation via googledrive::drive_auth()
}
