% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{run_gd_oauth}
\alias{run_gd_oauth}
\title{Runs google drive authorisation via googledrive::drive_auth() and saves credentials}
\usage{
run_gd_oauth(credential_name = "gd-credentials.rds")
}
\description{
Runs google drive authorisation via googledrive::drive_auth() and saves credentials
}
