% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{skip_test_if_not_possible}
\alias{skip_test_if_not_possible}
\title{skips test in testthat evaluation if requirement are not met. The function tests for credentials python modules and test files on google drive}
\usage{
skip_test_if_not_possible()
}
\description{
skips test in testthat evaluation if requirement are not met. The function tests for credentials python modules and test files on google drive
}
