% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{test_gdal_installation}
\alias{test_gdal_installation}
\title{test local gdal installation for vir-env workaround}
\usage{
test_gdal_installation()
}
\description{
test local gdal installation for vir-env workaround
}
