% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{test_import_ee_gdal_virtual}
\alias{test_import_ee_gdal_virtual}
\title{test import of gdal and ee for virtual env}
\usage{
test_import_ee_gdal_virtual()
}
\description{
test import of gdal and ee for virtual env
}
