% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload2ft}
\alias{upload2ft}
\title{upload vector data as fusion table and parse file to allow large uploads}
\usage{
upload2ft(path2file, fileName)
}
\arguments{
\item{path2file}{Path to vector data}

\item{fileName}{Name of fusion table in google drive}
}
\description{
upload vector data as fusion table and parse file to allow large uploads
}
