% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_as_ft}
\alias{upload_as_ft}
\title{upload vector data as fusion table and parse file to allow large uploads}
\usage{
upload_as_ft(file_path, fileName)
}
\arguments{
\item{file_path}{Path to vector data}

\item{fileName}{Name of fusion table in google drive}
}
\description{
upload vector data as fusion table and parse file to allow large uploads
}
