% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{upload_data_ft}
\alias{upload_data_ft}
\title{upload as fusion tables !!!!!!!!old version!!!!!!!!!!!!!!}
\usage{
upload_data_ft(path_file)
}
\arguments{
\item{path_file}{path of file to upload}
}
\value{
if 0
}
\description{
upload as fusion tables !!!!!!!!old version!!!!!!!!!!!!!!
}
