% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_data_old}
\alias{upload_data_old}
\title{old upload vector data and return fusion table ID}
\usage{
upload_data_old(verbose = T, target)
}
\arguments{
\item{verbose}{specifies weather information is about the process is printed to the console}

\item{target}{path to vector data to be uploaded}
}
\value{
Fusion table ID
}
\description{
old upload vector data and return fusion table ID
}
