% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{use_env}
\alias{use_env}
\title{Use a specific virtual or conda environment dependent on the operating system}
\usage{
use_env(env_name = "r-reticulate")
}
\description{
Use a specific virtual or conda environment dependent on the operating system
}
