% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{wait_for_file_on_drive}
\alias{wait_for_file_on_drive}
\title{wait_for_file_on_drive}
\usage{
wait_for_file_on_drive(filename, verbose = T)
}
\arguments{
\item{filename}{name of the file to scan drive for}
}
\description{
wait_for_file_on_drive
}
