% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{download_data_waiting}
\usage{
download_data(filename, path = getwd(), clear = T, verbose = T)
}
\arguments{
\item{path}{The local path where the file should be stored, default is working directory.}

\item{clear}{If the file should be removed from Google Drive after the download.}

\item{info}{Output of the get_data function.}
}
\value{
nothing
}
\description{
download_data_waiting
}
