% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{download_data_waiting}
\usage{
download_data(ee_response, clear = T, verbose = T, temp_path)
}
\arguments{
\item{ee_response}{Output of the request_data function.}

\item{clear}{If the file should be removed from Google Drive after the download.}
}
\value{
nothing
}
\description{
download_data_waiting
}
