% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_grab.R
\name{ee_grab}
\alias{ee_grab}
\title{Request and import data from Earth Engine}
\usage{
ee_grab(target = NULL, products = NULL, verbose = T)
}
\arguments{
\item{target}{A \code{character} path to a local geo-file that should be used as a target (.shp, .geojson, .kml). If the file is already uploaded, the upload is skipped.}

\item{products}{A \code{list} of \code{create_image_product()} or \code{create_collection_product()} functions which specify the requested data products. Multiple functions are passed inside a \code{list}, while a single function can be passed directly.}

\item{verbose}{A \code{logical}, whether to inform the user about the processing state of the data products. Default is set to \code{True}.}
}
\value{
Object of class \code{sf}. The \code{ee_grab()} returns the target file with the bands of the requested data products added as columns.
}
\description{
The \code{ee_grab()} function takes a \code{target} and \code{products}, generated by \code{create_image_product()} and \code{create_collection_product()} to request and import the data from the Earth Engine servers to R.
}
\details{
The \code{ee_grab()} processing runs in 4 steps:

1. Upload - The target is uploaded to Google Drive.

2. Request - The data products are requested from Google Earth Engine and exported to Google Drive.

3. Download -  The data products are downloaded from Drive.

4. Import - The data is imported to R and merged.
}
\examples{
\dontrun{
# Request a srtm image data product to get topographic data.
# Grab the spatial mean of the elevation band in the polygons of your target. The calculation are based on a 100 meter scale, which means that the original SRTM data product is resampled to 100 * 100 meter Pixel size.

srtm_data <- ee_grab(target = system.file("data/territories.shp", package = "earthEngineGrabR"),
                    products = create_image_product(productID = "CGIAR/SRTM90_V4", 
                                                    spatialReducer = "mean", 
                                                    scale = 100, 
                                                    bands = "elevation"
                                                    )
                   )
                                                    
# Request a chirps collection data product to get precipitation data.
# Grab the yearly precipitation sum for 2016 and get the spatial mean in the polygons of your target.

chirps_data <- ee_grab(target = system.file("data/territories.shp", package = "earthEngineGrabR"),
                      products = create_collection_product(productID = "UCSB-CHG/CHIRPS/DAILY",
                                                           spatialReducer = "mean",
                                                           temporalReducer = "sum", 
                                                           timeStart = "2016-01-01",
                                                           timeEnd = "2016-12-31", 
                                                           scale = 200
                                                           )
                     )                                                    
                                                    
                                                                                                                                                                                                                                                              
}

}
