% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_grab.R
\name{ee_grab}
\alias{ee_grab}
\title{ee_grab}
\usage{
ee_grab(target = NULL, outputFormat = "GeoJSON", resolution = 100,
  products = list(eeProduct_modis_treeCover()), verbose = T)
}
\arguments{
\item{target}{A path to a local geofile, if file is already uploaded, the upload is skipped.}

\item{outputFormat}{A string specifying the output format: CSV, GeoJSON, KML or KMZ.}

\item{resolution}{Resolution of the dataproducts.}

\item{products}{List of dataproduct functions starting with eeProduct}

\item{verbose}{if true, prints messages about the state of processing}
}
\value{
Object of class sf.
}
\description{
ee_grab
}
