% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{eeProduct_oxford_accessibility}
\alias{eeProduct_oxford_accessibility}
\title{oxford_accessibility}
\usage{
eeProduct_oxford_accessibility(spatialReducer = "mean")
}
\arguments{
\item{spatialReducer}{Reducer to spatially aggregate all dataproducts in each geometry of the feature, can be: mean, median or mode)}
}
\value{
depend on output
}
\description{
oxford_accessibility
}
