% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{exec_auth_new_window}
\alias{exec_auth_new_window}
\title{Execute command in new terminal window for all operating systems}
\usage{
exec_auth_new_window(command, gnome = T, credential_path, credential_name)
}
\description{
Execute command in new terminal window for all operating systems
}
