% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gd_auth}
\alias{gd_auth}
\title{retreves credentials and runs google drive authorisation via googledrive::drive_auth()}
\usage{
gd_auth(credential_name = "gd-credentials.rds")
}
\description{
retreves credentials and runs google drive authorisation via googledrive::drive_auth()
}
