% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data
calls get_data_image or get_data_collections, dependent on the info object}
\usage{
get_data(info)
}
\arguments{
\item{info}{Data frame information generated gy ee_grab()}

\item{data_type}{either ImageCollection of Image,}
}
\description{
get_data
calls get_data_image or get_data_collections, dependent on the info object
}
