% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(year_start = 2000, year_end = 2000, time_reducer = "mean",
  asset_path = NULL, output = "GeoJSON", spatial_reducer = "mean",
  scale = 100, name = "example", products = list("chirps_precipitation",
  "jrc_permanentWater", "modis_treeCover", "modis_nonTreeVegetation",
  "modis_nonVegetated", "srtm_elevation", "srtm_slope", "modis_quality"))
}
\arguments{
\item{year_start}{Integer to spedify the beginning of timeperiod to reduce over.}

\item{year_end}{Integer to spedify the end of timeperiod to reduce over.}

\item{time_reducer}{Reducer to aggregate data over time, can be: mean, median or mode}

\item{asset_path}{A string path to earth engine asset}

\item{output}{A string specifying the output format: CSV, GeoJSON, KML or KMZ}

\item{spatial_reducer}{Reducer to spatially aggregate data, can be: mean, median or mode}

\item{scale}{A scale spedifying the basis of analysis}

\item{products}{list("chirps_precipitation","jrc_permanentWater","modis_treeCover","modis_nonTreeVegetation","modis_nonVegetated","srtm_elevation",srtm_slope", "modis_quality")}
}
\value{
depend on output
}
\description{
get_data
}
