% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{get_data_info}
\alias{get_data_info}
\title{get_data_info
retreves info with a given product ID over earthEngine}
\usage{
get_data_info(productID)
}
\arguments{
\item{productID}{String that speciefies a data products in ee}
}
\description{
get_data_info
retreves info with a given product ID over earthEngine
}
