% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{import_data}
\alias{import_data}
\title{import data}
\usage{
import_data(product_list, verbose = T, temp_path)
}
\arguments{
\item{verbose}{If true, messages reporting the processing state are printed.}

\item{productList}{List of products files produced in the ee_grab function}
}
\value{
nothing
}
\description{
import data
}
