% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oAuth.R
\name{test_credentials}
\alias{test_credentials}
\title{Test if credentials can be found in the default location and raises an error message of not.}
\usage{
test_credentials(with_error = F)
}
\arguments{
\item{with_error}{A logical weather to raise an informative error in case of missing credentials.}
}
\description{
Test if credentials can be found in the default location and raises an error message of not.
}
