% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthEngineGrabR.R
\name{upload_data}
\alias{upload_data}
\title{deletes credentials to re initialize}
\usage{
upload_data(verbose = T, target)
}
\arguments{
\item{verbose}{specifies weather information is about the process is printed to the console}

\item{target}{path to vector data to be uploaded}
}
\description{
deletes credentials to re initialize
}
