% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_data}
\alias{upload_data}
\title{upload vector data and return fusion table ID}
\usage{
upload_data(targetArea, verbose = T)
}
\arguments{
\item{targetArea}{path to vector data to be uploaded}

\item{verbose}{specifies weather information is about the process is printed to the console}
}
\value{
Fusion table ID
}
\description{
upload vector data and return fusion table ID
}
