% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE2R.R
\name{validate_shapefile}
\alias{validate_shapefile}
\title{validate_shapefile}
\usage{
validate_shapefile(ft_id)
}
\arguments{
\item{ft_id}{id of fusion table}
}
\value{
if shapefile is not valid, returns error
}
\description{
validate_shapefile
}
